+++ python/GafferUI/NumericWidget.py
@@ -36,6 +36,10 @@
 ##########################################################################
 
 import math
+import re
+import operator
+
+import six
 
 import IECore
 
@@ -75,7 +79,10 @@ class NumericWidget( GafferUI.TextWidget ) :
 
 	def getValue( self ) :
 
-		return self.__numericType( self.getText() )
+		# Call `fixup()` so that we can always return a valid value,
+		# even if the user is part way through entering an expression.
+		value = self._qtWidget().validator().fixup( self.getText() )
+		return self.__numericType( value )
 
 	## A signal emitted whenever the value has been changed and the user would expect
 	# to see that change reflected in whatever the field controls. Slots should have
@@ -273,15 +280,7 @@ class NumericWidget( GafferUI.TextWidget ) :
 		if self.__numericType is not numericType :
 
 			self.__numericType = numericType
-
-			if self.__numericType is int :
-				validator = QtGui.QIntValidator( self._qtWidget() )
-			else :
-				validator = QtGui.QDoubleValidator( self._qtWidget() )
-				validator.setDecimals( 4 )
-				validator.setNotation( QtGui.QDoubleValidator.StandardNotation )
-
-			self._qtWidget().setValidator( validator )
+			self._qtWidget().setValidator( _ExpressionValidator( self._qtWidget(), self.__numericType ) )
 
 		# update our textual value
 		text = self.__valueToString( value )
@@ -304,3 +303,94 @@ class NumericWidget( GafferUI.TextWidget ) :
 			return
 
 		signal( self, reason )
+
+# A basic validator/evaluator that supports simple maths
+# operators [+-/*] along with standard number validation, eg:
+#   2 + 3
+#   4.4 / 2
+# Qt will call validate/fixup as part of the edit cycle.
+class _ExpressionValidator( QtGui.QValidator ) :
+
+	def __init__( self, parent, numericType ) :
+
+		QtGui.QValidator.__init__( self, parent )
+
+		self.__numericType = numericType
+		if self.__numericType is int :
+			operand = r"-?[0-9]*"
+		else :
+			operand = r"-?[0-9]*\.?[0-9]{0,4}"
+
+		# Captures `( operand1, operator, operand2 )``, with the latter
+		# two being optional. Note that the regex for the operands is
+		# deliberately loose and accepts incomplete text such as "", "-"
+		# and ".". This is dealt with in `validate()` and `fixup()`.
+		self.__expression = re.compile(
+			r"^\s*({operand})\s*(?:([-+*/%])\s*({operand}))?\s*$".format(
+				operand = operand
+			)
+		)
+
+	def fixup( self, text ) :
+
+		match = re.match( self.__expression, text )
+		try :
+			operand1 = self.__numericType( match.group( 1 ) )
+		except :
+			return "0"
+
+		try :
+			operand2 = self.__numericType( match.group( 3 ) )
+		except :
+			# Incomplete expression, return first number
+			return match.group( 1 )
+
+		# Evaluate expression
+
+		op = {
+			"/" : operator.floordiv if self.__numericType is int else operator.truediv,
+			"*" : operator.mul,
+			"+" : operator.add,
+			"-" : operator.sub,
+			"%" : operator.mod
+		}[match.group(2)]
+
+		try :
+			return str( op( operand1, operand2 ) )
+		except ZeroDivisionError :
+			return match.group( 1 )
+
+	def validate( self, text, pos ) :
+
+		# Work around https://bugreports.qt.io/browse/PYSIDE-106. Because
+		# that prevents `fixup()` from working, it stops Qt automatically
+		# fixing `Intermediate` input when editing finishes, which in turn stops
+		# `QLineEdit.editingFinished` from being emitted. We return `Acceptable`
+		# instead of `Intermediate` and then apply `fixup()` ourselves via
+		# `__editingFinished()` and `getValue()`.
+		## \todo Remove once we've moved to GafferHQ/dependencies 2.0.0, which
+		# contains the relevant PySide fix. Include a grace period for folks
+		# building with their own older dependencies.
+		Intermediate = QtGui.QValidator.Acceptable
+
+		match = re.match( self.__expression, text )
+		if match is None :
+			return QtGui.QValidator.Invalid, text, pos
+
+		# Check first operand
+		try :
+			self.__numericType( match.group( 1 ) )
+		except :
+			return (
+				Intermediate if not match.group( 2 ) else QtGui.QValidator.Invalid,
+				text, pos
+			)
+
+		# If operator is present, check second operand
+		if match.group( 2 ) :
+			try :
+				self.__numericType( match.group( 3 ) )
+			except :
+				return Intermediate, text, pos
+
+		return QtGui.QValidator.Acceptable, text, pos